perf(geometry): the focal vote takes f32 positions and scores its epipolar cell in f32
The interface stops pretending: member_positions is f32 through the
core entries, the bindings and MatchesObservations (which now borrows
the file's pairs with no widening) -- the file stores f32, keypoint
noise sits three orders above f32 resolution, and the widened f64
carried zero information. The pair-table read widens exactly, all f64
arithmetic is unchanged, and a 42-capture fleet probe deep-compares
bit-identical.

The epipolar cell's residual loop then computes in f32 by default:
rays narrowed per candidate focal (the f64 buffers hold the widened
narrowed value, so elimination rows and kabsch consume the same ray
the residual measured), an 8-lane AVX2 kernel with single-precision
sqrt/div and an op-for-op scalar twin. The residual is a directly
computed sine ratio, well conditioned at small values; measured, the
kernel runs 2.7-3.2x and the escalated vote 12-19% faster, with every
fleet verdict, confirmation and escalation invariant, consensus focals
within 7e-6, and the only discrete movement +-1 informative mass on
two wide-margin entries double-backstopped by the confirmation rule.
SFMTOOL_FOCAL_VOTE_F64_EPI is the diagnostic restore, in the
convention of the other two.

The rotation cell stays f64: the f32-viable cross-product form is
measured performance-neutral (the extra cross costs what the lane
doubling saves), and acos of an f32 dot is disqualified outright
(median 38% error in the inlier band). The measured arm ships gated
off (SFMTOOL_FOCAL_VOTE_F32_ROT) as the record of that result.

Also: the residual dispatch's consecutiveness test dropped from an
O(n) scan per call to O(1) (callers pass strictly increasing indexes),
worth 9-11% of escalated wall by itself, bit-identically; and
SFMTOOL_PROFILE=1 phase instrumentation (focal_vote/prof.rs) stays,
env-gated, in the cluster_refine precedent.

+++ crates/sfmtool-core/src/geometry/estimate_intrinsics.rs
@@ -253,7 +253,7 @@ pub struct IntrinsicsEstimate {
 /// // One two-member cluster, seen in images 0 and 1.
 /// let cluster_starts = [0u32, 2];
 /// let member_images = [0u32, 1];
-/// let member_positions = [[120.0, 240.0], [131.5, 238.0]];
+/// let member_positions = [[120.0f32, 240.0], [131.5, 238.0]];
 /// let estimate = estimate_intrinsics(
 ///     &cluster_starts,
 ///     &member_images,
@@ -269,7 +269,7 @@ pub struct IntrinsicsEstimate {
 pub fn estimate_intrinsics(
     cluster_starts: &[u32],
     member_images: &[u32],
-    member_positions: &[[f64; 2]],
+    member_positions: &[[f32; 2]],
     width: u32,
     height: u32,
     options: &IntrinsicsOptions,
